bugfix: fix pipeline bug

+++ deploy.sh
@@ -10,12 +10,12 @@ fi
 # Check if nginx is running
 if [ "$(docker container ls --filter name=nginx -q)" ]; then
   # Check which app (blue or green) is currently running
-  if [ "$(docker container ls --filter name=demo_app_blue -q)" ]; then
+  if [ "$(docker container ls -a --filter name=demo_app_blue -q)" ]; then
     # Run the green app and stop the blue app
     docker-compose -f docker-compose.yml up -d --no-recreate
     docker container rm -f demo_app_blue
     echo "New version is demo app green"
-  elif [ "$(docker container ls --filter name=demo_app_green -q)" ]; then
+  elif [ "$(docker container ls -a --filter name=demo_app_green -q)" ]; then
     # Run the blue app and stop the green app
     docker-compose -f docker-compose.yml up -d --no-recreate
     docker container rm -f demo_app_green
